fastapi/requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.6.2
asgi-correlation-id==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.3
cryptography==41.0.2
Deprecated==1.2.14
dnspython==2.4.1
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.110.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.0
httptools==0.6.1
httpx==0.24.0
idna==3.4
importlib-metadata==6.11.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mypy==1.7.1
mypy-extensions==1.0.0
opentelemetry-api==1.23.0
opentelemetry-distro==0.44b0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-instrumentation-logging==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.8.10
packaging==23.1
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.4.0
prometheus_client==0.20.0
protobuf==4.25.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==2.1.1
pydantic-settings==2.0.3
pydantic_core==2.4.0
pyinstrument==4.6.1
pytest==8.0.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
ruff==0.1.6
sentry-sdk==1.40.6
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.36.3
starlette_exporter==0.21.0
tomli==2.0.1
typing_extensions==4.10.0
ujson==5.7.0
urllib3==2.2.1
uvicorn==0.21.1
uvloop==0.19.0
watchfiles==0.19.0
websockets==11.0.1
wrapt==1.16.0
zipp==3.18.1
